requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.8
django-cors-headers==4.3.1
django-ratelimit==4.1.0
django-service-objects==0.7.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
PyJWT==1.7.1
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.9.0
